dpaste: use a random PIN instead of random hash
The whole point of a pastebin is the ability to easily
share/copy the "link" of the pasted text, easier than
the text to copy itself.

A full hash is too long to be copied by a human being.
By using a random 32-bits PIN it is possible
for a human being to transmit the text between computers
without somehow transmitting the hash.

Currently the 32-bits is in hex form, but
could be made base64 or some other tighter format.